+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ﻿# Changelog
 
+## [1.0.1] -2025-12-14
+
+- Added global production rate adjustment for extractors, processors, and unique factories (Mod Default).
+- Added compatibility support for the Campus Industries Housing Mod.
+- Added support for the Fishing Harbor building.
+- Fixed an issue where the entry button was still displayed on certain unavailable industrial buildings.
+- Fixed a potential deserialization issue that could prevent the mod from working.
+
 ## [1.0.0] - 2025-12-08
+
 - Updated to support game version 1.20.1
 - Updated to CSLModsCommon framework 1.0
 - This update automatically resets all settings.
@@ -9,18 +18,22 @@
 - Added mod default values, game default values, custom options to modify industrial building properties.
 
 ## [0.9.3] - 2024-09-07
+
 - Fixed serialization exception issues.
 
 ## [0.9.2] - 2024-08-31
+
 - Code cleanup.
 - Updated to CSShared framework.
 - Updated localization.
 
 ## [0.9.1] - 2024-07-20
+
 - Code cleanup.
 - Fix potential issues.
 
 ## [0.9.0] - 2023-06-28
+
 - Updated to support game version 1.17.1
 - Added control panel to provide more convenient control entry.
 - Added options to control the number of trucks in industrial facilities.
@@ -28,6 +41,7 @@
 - Fixed exception throws that can be caused by Customize It Extended.
 
 ## [0.8.0] - 2023-05-23
+
 - Updated to support game version 1.17.0
 - Added Korean translation.
 - Added Czech translation.
@@ -36,40 +50,49 @@
 - Optimized input/output buffer size control logic.
 
 ## [0.7.1.1] - 2023-04-30
+
 [FIX]Fixed an issue with unadjustable processing rates.
 
 ## [0.7.1] - 2023-03-22
+
 [ADD]Added advanced option for reset mod config.
 
 ## [0.7.0] - 2023-03-11
+
 - Fixed incorrect number of vehicles in the unique factory.
 - Updated framework and UI style.
 
 ## [0.6.0] - 2023-03-08
+
 - Added Extracting facility and processing facility output rate regulation.
 - Optimized rebinding prefab speed.
 - Fixed some farm buildings overemployment issues.
 
 ## [0.5.0] - 2023-02-11
+
 - Added tooltip box mode option to compare the original value to the corrected value.
 - Fixed tooltip box for some buildings still not showing the corrected value.
 - Fixed worker level up requirements are too high issues.
 
 ## [0.4.0] - 2023-02-05
+
 - Fixed incorrect maintenance costs for unique factories.
 - Fixed issues where hovering over a building pops up tooltip box with incorrect construction costs, maintenance costs, vehicle counts, and workspace.
 
 ## [0.3.0] - 2023-01-17
+
 - Fixed an issue where non-vanilla buildings were not spawns vehicles correctly.
 - Fixed an issue where work space was incorrect.
 - Fixed compatibility issues with IMT and IPT mod.
 - Fixed the exception throwing issues.
 
 ## [0.2.0] - 2023-01-06
+
 - Update buffer recognition and rebinding logic.
 - Fix some processing buildings appear not enough buyer or not enough raw materials issues.
 
 ## [0.1.0] - 2023-01-02
+
 - Fixed incorrect display of vehicle count.
 - Using new algorithms to rebalances Industries.
 - Using the latest Harmony API.